+++ packages/sdk/src/metrics/get-rewards-info.ts
@@ -4,11 +4,12 @@ import { getApiConfig } from '../common/api-config';
 import { IEnvParam } from '../common/parameters';
 
 type Response = {
-  type: 'BALANCE_TYPE_DEFI';
+  type: 'BALANCE_TYPE_DEFI' | 'BALANCE_TYPE_HOLDING';
   total_lbtc_balance: number;
   total_lbtc_balance_cost: number;
   total_rewards: number;
   total_rewards_cost: number;
+  chains: { lbtc_balance: number; chain: string }[];
 };
 
 /**
@@ -22,7 +23,7 @@ export async function getRewardsInfo({
 }: { account: Address } & IEnvParam) {
   const { baseApiUrl } = getApiConfig(env);
 
-  const url = `${baseApiUrl}/api/v1/rewards/${account}`;
+  const url = `${baseApiUrl}/api/v1/rewards/summary/${account}`;
   const { data } = await axios.get<Response>(url);
 
   return {
@@ -30,5 +31,24 @@ export async function getRewardsInfo({
     totalLbtcBalance: data.total_lbtc_balance,
     totalRewards: data.total_rewards,
     totalRewardsCost: data.total_rewards_cost,
+    chains: data.chains?.map(cd => ({
+      balance: cd.lbtc_balance,
+      chain: cd.chain, // TODO: Map this to chain id.
+    })),
   };
 }
+// Response example:
+//
+// {
+//   "type":"BALANCE_TYPE_HOLDING",
+//   "total_lbtc_balance":0.00030997,
+//   "total_lbtc_balance_cost":36.906438563500004,
+//   "total_rewards":3.278178613395223e-8,
+//   "total_rewards_cost":0.0039031486142352624,
+//   "chains":[
+//     {
+//       "lbtc_balance":0.00030997,
+//       "chain":"ethereum"
+//     }
+//   ]
+// }